\name{print.AcrossTic}
\alias{print.AcrossTic}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print method for AcrossTic objects
}
\description{
Print some attributes of an AcrossTic object to the screen.
}
\usage{
\method{print}{AcrossTic} (x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
AcrossTic item (output from \code{\link{rRegMatch}}) to be printed
}
  \item{\dots}{
Other arguments, currently ignored.
}
}

\value{
None.
}

\author{
Sam Buttrey
}
