\name{print.AcrossTicPtest}
\alias{print.AcrossTicPtest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print output of AcrossTic permutation test
}
\description{
Print the output of a permutation test on an AcrossTic object (see \code{\link{ptest}})
}
\usage{
\method{print}{AcrossTicPtest} (x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Object of class AcrossTicPtest
}
  \item{\dots}{
Other arguments, currently ignored.
}
}
\details{
The output from \code{\link{ptest}} has class AcrossTicPtest. This function prints such an object.
}
\value{
None
}
\author{
Sam Buttrey
}
